Added note userId authentication

+++ backend/src/controllers/notes.ts
@@ -2,11 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+    
     try {
+        assertIsDefined(authenticatedUserId);
         
-        const notes = await NoteModel.find().exec();
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -15,8 +19,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
-
+    const authenticatedUserId = req.session.userId;
+    
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note ID");
         }
@@ -26,6 +32,8 @@ export const getNote: RequestHandler = async (req, res, next) => {
             throw createHttpError(404, "Note not found");
         }
 
+        if (!note.userId.equals(authenticatedUserId)) throw createHttpError(401,"You don't have access to this note");
+        
         res.status(200).json(note);
     } catch (error) {
         next(error);
@@ -40,13 +48,17 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
+    
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!title) {
             throw createHttpError(400, "Note title is required")
         }
 
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -70,8 +82,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note ID");
         }
@@ -86,6 +100,7 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
             throw createHttpError(404, "Note not found");
         }
 
+        if (!note.userId.equals(authenticatedUserId)) throw createHttpError(401,"You don't have access to this note");
 
         note.title = newTitle;
         note.text = newText;
@@ -99,8 +114,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note ID");
         }
@@ -111,6 +128,8 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
             throw createHttpError(404, "Note note found");
 
         }
+
+        if (!note.userId.equals(authenticatedUserId)) throw createHttpError(401,"You don't have access to this note");
         
         NoteModel.findByIdAndRemove(noteId).exec();
